# docs/flatmates-openapi.yaml
openapi: 3.1.0
info:
  title: 360 Flatmates API
  description: |
    Complete API for the 360 Flatmates web platform — a swipe-based flatmate-finding app for India's young professionals.

    ## Authentication
    All endpoints require authentication via Bearer token in the `Authorization` header, unless marked as public.
    Tokens are obtained via Supabase Auth (Phone OTP + Password).

    ## User Modes
    - **room_poster** — Has a flat, wants to rent out a spare room
    - **co_hunter** — Looking for people to flat-search with
    - **open_to_both** — Flexible, open to either

    ## Base URL
    `/api/v1`
  version: 1.0.0
  contact:
    name: 360 Ghar
    url: https://360ghar.com
  license:
    name: Confidential
    url: https://360ghar.com
servers:
  - url: https://api.360ghar.com/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local Development
...
paths:
...
  /users/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: Validate the authenticated user exists and return basic user info.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/location:
    get:
      tags: [Auth]
      summary: Get user location
      description: Returns the current user's location data.
      operationId: getUserLocation
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLocation'

  # ===================== BOOTSTRAP =====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        phone:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [user, admin, agent]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        preferences:
          type: object
          additionalProperties: true
    UserLocation:
      type: object
      properties:
        city:
          type: string
        locality:
          type: string
        lat:
          type: number
          format: float
        lng:
          type: number
          format: float

    # ===================== BOOTSTRAP SCHEMAS =====================
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
        error_code:
          type: string
          description: Optional machine-readable error code